+++ desktop/js/update.js
@@ -314,7 +314,7 @@ function addUpdate(_update) {
   tr += '<td style="width:160px;"><span class="label label-primary" data-l1key="remoteVersion">'+_remoteVersion+'</span></td>';
   tr += '<td style="width:180px;">';
   if (_update.type != 'core') {
-    tr += '<input type="checkbox" class="updateAttr" data-l1key="configuration" data-l2key="doNotUpdate" title="Sauvegarder pour conserver les modications"><span>{{Ne pas mettre à jour}}</span>';
+    tr += '<input type="checkbox" class="updateAttr" data-l1key="configuration" data-l2key="doNotUpdate" title="{{Sauvegarder pour conserver les modications}}"><span>{{Ne pas mettre à jour}}</span>';
   }
   tr += '</td>';
   tr += '<td style="width:350px;">';