+++ desktop/php/interact.php
@@ -197,7 +197,7 @@ $optionMaxSize = 15;
 							<textarea class="form-control interactAttr ta_autosize" type="text" data-l1key="reply" placeholder=""></textarea>
 						</div>
 						<div class="col-sm-1">
-							<a class="btn btn-default cursor listEquipementInfoReply input-sm" title="Rechercher une commande"><i class="fas fa-list-alt "></i></a>
+							<a class="btn btn-default cursor listEquipementInfoReply input-sm" title="{{Rechercher une commande}}"><i class="fas fa-list-alt "></i></a>
 						</div>
 					</div>
 					<div class="form-group">